requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dotenv==1.0.1
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.7.0 